Agrega manejo de archivos con S3

+++ src/main/java/com/ayds/Cloudmerce/model/dto/cart/CartDTO.java
@@ -24,7 +24,7 @@ public class CartDTO {
     @NonNull @Positive
     private BigDecimal tax;
     @NonNull @Positive
-    private Integer paymentMethodId;
+    private Long paymentMethodId;
     @NonNull @Positive
     private Integer statusId;
     @NonNull
